moves some helpers out from the component
